Different file names for C and C++: include files use hpp.
