+++ MQTTClient-CPP/samples/CMakeLists.txt
@@ -1,5 +1,5 @@
 #*******************************************************************************
-#  Copyright (c) 2017, 2018 IBM Corp.
+#  Copyright (c) 2017, 2023 IBM Corp.
 #
 #  All rights reserved. This program and the accompanying materials
 #  are made available under the terms of the Eclipse Public License v1.0
@@ -14,25 +14,4 @@
 #     Ian Craggs - initial version
 #*******************************************************************************/
 
-# Samples
-
-add_executable(
-  hello
-  hello.cpp
-)
-target_include_directories(hello PRIVATE "../../src" "../../src/linux")
-target_link_libraries(hello MQTTPacketClient)
-
-add_executable(
-  hello_class
-  hello_class.cpp
-)
-target_include_directories(hello_class PRIVATE "../../src" "../../src/linux")
-target_link_libraries(hello_class MQTTPacketClient)
-
-add_executable(
-  stdoutsub
-  stdoutsub.cpp
-)
-target_include_directories(stdoutsub PRIVATE "../../src" "../../src/linux")
-target_link_libraries(stdoutsub paho-embed-mqtt3c)
+ADD_SUBDIRECTORY(linux)
