+++ MQTTClient-CPP/src/arduino/Countdown.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2014, 2017 IBM Corp.
+ * Copyright (c) 2014, 2023 IBM Corp.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -12,54 +12,46 @@
  *
  * Contributors:
  *    Ian Craggs - initial API and implementation and/or initial documentation
- *    Ian Craggs - change Timer member initialization to avoid copy constructor
  *******************************************************************************/
 
-#if !defined(MQTT_MBED_H)
-#define MQTT_MBED_H
-
-#include "mbed.h"
+#if !defined(COUNTDOWN_H)
+#define COUNTDOWN_H
 
 class Countdown
 {
 public:
-    Countdown() : t()
-    {
-
+    Countdown()
+    {  
+		interval_end_ms = 0L;
     }
-
-    Countdown(int ms) : t()
+    
+    Countdown(int ms)
     {
-        countdown_ms(ms);
+        countdown_ms(ms);   
     }
-
-
+    
     bool expired()
     {
-        return t.read_ms() >= interval_end_ms;
+        return (interval_end_ms > 0L) && (millis() >= interval_end_ms);
     }
-
-    void countdown_ms(unsigned long ms)
+    
+    void countdown_ms(unsigned long ms)  
     {
-        t.stop();
-        interval_end_ms = ms;
-        t.reset();
-        t.start();
+        interval_end_ms = millis() + ms;
     }
-
+    
     void countdown(int seconds)
     {
         countdown_ms((unsigned long)seconds * 1000L);
     }
-
+    
     int left_ms()
     {
-        return interval_end_ms - t.read_ms();
+        return interval_end_ms - millis();
     }
-
+    
 private:
-    Timer t;
-    unsigned long interval_end_ms;
+    unsigned long interval_end_ms; 
 };
 
 #endif
